+++ package.json
@@ -1,45 +1,45 @@
 {
-    "name": "src",
-    "version": "0.1.0",
-    "private": true,
-    "scripts": {
-        "dev": "next dev --turbopack",
-        "build": "next build",
-        "start": "next start",
-        "lint": "next lint"
-    },
-    "dependencies": {
-        "@iconify-json/simple-icons": "^1.2.49",
-        "@radix-ui/react-collapsible": "^1.1.12",
-        "@radix-ui/react-dialog": "^1.1.15",
-        "@radix-ui/react-navigation-menu": "^1.2.14",
-        "@radix-ui/react-slot": "^1.2.3",
-        "class-variance-authority": "^0.7.1",
-        "clsx": "^2.1.1",
-        "embla-carousel-react": "^8.6.0",
-        "lucide-react": "^0.542.0",
-        "next": "15.4.2",
-        "react": "19.1.0",
-        "react-dom": "19.1.0",
-        "slick-carousel": "^1.8.1",
-        "tailwind-merge": "^3.3.1"
-    },
-    "devDependencies": {
-        "@eslint/eslintrc": "^3",
-        "@iconify-icon/react": "^3.0.0",
-        "@iconify-json/bi": "^1.2.4",
-        "@iconify-json/material-symbols": "^1.2.33",
-        "@iconify/tailwind4": "^1.0.6",
-        "@tailwindcss/postcss": "^4",
-        "@types/node": "^20",
-        "@types/react": "^19",
-        "@types/react-dom": "^19",
-        "@types/react-slick": "^0.23.13",
-        "eslint": "^9",
-        "eslint-config-next": "15.4.2",
-        "react-slick": "^0.30.3",
-        "tailwindcss": "^4",
-        "tw-animate-css": "^1.3.7",
-        "typescript": "^5"
-    }
+  "name": "src",
+  "version": "0.1.0",
+  "private": true,
+  "scripts": {
+    "dev": "next dev",
+    "build": "next build",
+    "start": "next start",
+    "lint": "next lint"
+  },
+  "dependencies": {
+    "@iconify-json/simple-icons": "^1.2.49",
+    "@radix-ui/react-collapsible": "^1.1.12",
+    "@radix-ui/react-dialog": "^1.1.15",
+    "@radix-ui/react-navigation-menu": "^1.2.14",
+    "@radix-ui/react-slot": "^1.2.3",
+    "class-variance-authority": "^0.7.1",
+    "clsx": "^2.1.1",
+    "embla-carousel-react": "^8.6.0",
+    "lucide-react": "^0.542.0",
+    "next": "15.4.2",
+    "react": "19.1.0",
+    "react-dom": "19.1.0",
+    "slick-carousel": "^1.8.1",
+    "tailwind-merge": "^3.3.1"
+  },
+  "devDependencies": {
+    "@eslint/eslintrc": "^3",
+    "@iconify-icon/react": "^3.0.0",
+    "@iconify-json/bi": "^1.2.4",
+    "@iconify-json/material-symbols": "^1.2.33",
+    "@iconify/tailwind4": "^1.0.6",
+    "@tailwindcss/postcss": "^4",
+    "@types/node": "^20",
+    "@types/react": "^19",
+    "@types/react-dom": "^19",
+    "@types/react-slick": "^0.23.13",
+    "eslint": "^9",
+    "eslint-config-next": "15.4.2",
+    "react-slick": "^0.30.3",
+    "tailwindcss": "^4",
+    "tw-animate-css": "^1.3.7",
+    "typescript": "^5"
+  }
 }
